Core-Memory Partitioning Strategy
This implements a partitioning strategy that takes into account compute cores as a rough heuristic for model throughput. Even though it's not a super great heuristic, it will consistently distinguish GPU from CPU compute and will generally start the fastest-computing nodes first. It still partitions by memory, because there's actually no way not to do that without taking model size into account in the partitioning strategy

The motivation for prioritizing fastest throughput first is that the sequential nature of these models will necessitate that nodes later in the chain are waiting for nodes earlier in the chain to complete, so the computation will be held up by the slowest node disproportionately if it executes faster than earlier nodes. Note that we would need to take model size into account to maximize compute adjacency, as we should actually be trying to do inference with the minimum number of nodes possible if we want to optimize for speed. Still, this should be a small improvement