Throughput spillover partitioning strategy
This strategy builds on the throughput estimate from the core heuristic strategy, but modifies the way shards are allocated to allow maximizing model adjacency, trying to fit as much of the model on the fastest node in the network as is possible, then spilling over to the next fastest, until the model is fully hosted.

This breaks the guarantee that the model is distributed across all the nodes, but I actually think that's a good guarantee to stop making: It unnecessarily adds more network communication as a slow step across what is necessarily a sequential calculation as implemented. Some refactoring was done to enable this behavior